Separate light shadowmap and light rendering
